Import legacy WordPress Betheme Child files
This commits is authored by multiple contributors.

Contact Wikimedia Italia to discover the full list of these contributors
via this certified email address:

	wikimediaitalia@pec.it

Thanks to the BeTheme developers:
https://themeforest.net/item/betheme-responsive-multipurpose-wordpress-theme/7758048

Thanks to the NP solutions service provider who probably developed some patches:

	https://www.npsolutions.it/

Thanks to all the contributors that cannot be listed here for technical reasons but that,
again, you can discover contacting Wikimedia Italia.

This repository was initialized in order to track changes for this purpose:

https://phabricator.wikimedia.org/T267000